requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
cmdstanpy==1.1.0
contourpy==1.1.0
convertdate==2.4.0
cycler==0.11.0
Django==4.2.3
ephem==4.1.4
fonttools==4.41.1
holidays==0.29
idna==3.4
importlib-resources==6.0.0
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
LunarCalendar==0.0.9
MarkupSafe==2.1.3
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
plotly==5.15.0
prophet==1.1.4
PyMeeus==0.5.12
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sklearn==0.0.post7
sqlparse==0.4.4
tenacity==8.2.2
threadpoolctl==3.2.0
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.4
